library_management/requirements.txt
# Django Core
Django>=5.0

# Forms
django-crispy-forms>=2.0
crispy-bootstrap5>=2024.1

# Image Processing
Pillow>=10.0

# Environment Variables
python-dotenv>=1.0

# Security - Password Hashing (Django uses PBKDF2 by default, Argon2 recommended)
argon2-cffi>=23.1.0

# File Type Validation (Windows support)
python-magic-bin>=0.4.14


# Production Server (optional)
gunicorn>=21.0

# Database (optional - PostgreSQL)
# psycopg2-binary>=2.9

# Development/Testing
# bandit>=1.7  # Security linting
# safety>=2.3  # Dependency vulnerability scanning
